--- a/worker/backup_handler.go
+++ b/worker/backup_handler.go
@@ -100,19 +100,6 @@ type UriHandler interface {
 	ReadManifest(string, *Manifest) error
 }
 
-// getHandler returns a UriHandler for the URI scheme.
-func getHandler(scheme string, creds *x.MinioCredentials) UriHandler {
-	switch scheme {
-	case "file", "":
-		return &fileHandler{}
-	case "minio", "s3":
-		return &s3Handler{
-			creds: creds,
-		}
-	}
-	return nil
-}
-
 // NewUriHandler parses the requested URI and finds the corresponding UriHandler.
 // If the passed credentials are not nil, they will be used to override the
 // default credentials (only for backups to minio or S3).
@@ -139,12 +126,14 @@ func getHandler(scheme string, creds *x.MinioCredentials) UriHandler {
 //   file:///tmp/dgraph/backups
 //   /tmp/dgraph/backups?compress=gzip
 func NewUriHandler(uri *url.URL, creds *x.MinioCredentials) (UriHandler, error) {
-	h := getHandler(uri.Scheme, creds)
-	if h == nil {
-		return nil, errors.Errorf("Unable to handle url: %s", uri)
+	switch uri.Scheme {
+	case "file", "":
+		return &fileHandler{}, nil
+	case "minio", "s3":
+		return NewS3Handler(uri, creds)
 	}
+	return nil, errors.Errorf("Unable to handle url: %s", uri)
 
-	return h, nil
 }
 
 // loadFn is a function that will receive the current file being read.
@@ -162,8 +151,8 @@ func LoadBackup(location, backupId string, backupNum uint64, creds *x.MinioCrede
 		return LoadResult{Err: err}
 	}
 
-	h := getHandler(uri.Scheme, creds)
-	if h == nil {
+	h, err := NewUriHandler(uri, creds)
+	if err != nil {
 		return LoadResult{Err: errors.Errorf("Unsupported URI: %v", uri)}
 	}
 
@@ -178,9 +167,9 @@ func VerifyBackup(req *pb.RestoreRequest, creds *x.MinioCredentials, currentGrou
 		return err
 	}
 
-	h := getHandler(uri.Scheme, creds)
-	if h == nil {
-		return errors.Errorf("Unsupported URI: %v", uri)
+	h, err := NewUriHandler(uri, creds)
+	if err != nil {
+		return errors.Wrap(err, "VerifyBackup")
 	}
 
 	return h.Verify(uri, req, currentGroups)
@@ -193,9 +182,9 @@ func ListBackupManifests(l string, creds *x.MinioCredentials) (map[string]*Manif
 		return nil, err
 	}
 
-	h := getHandler(uri.Scheme, creds)
-	if h == nil {
-		return nil, errors.Errorf("Unsupported URI: %v", uri)
+	h, err := NewUriHandler(uri, creds)
+	if err != nil {
+		return nil, errors.Wrap(err, "ListBackupManifests")
 	}
 
 	paths, err := h.ListManifests(uri)