The list backups function was reading manifest files sequentially and
this is expensive. This PR improves the latency by parallelizing the
list backup calls.

For 10 backups stored in a single s3 bucket, list backups takes 31 seconds on master
```
dgraph lsbackup -l "s3:///amanbansal-test-bucket"  0.90s user 0.20s system 3% cpu 31.356 total
```
On this PR, it takes ~4.5 seconds
```
dgraph lsbackup -l "s3:///amanbansal-test-bucket"  0.74s user 0.08s system 18% cpu 4.566 total
```
<!--
Your title must be in the following format: topic(Area): Feature
Topic must be one of build|ci|docs|feat|fix|perf|refactor|chore|test

Sample Titles:
feat(Enterprise): Backups can now get credentials from IAM
fix(Query): Skipping floats that cannot be Marshalled in JSON
perf: [Breaking] json encoding is now 35% faster if SIMD is present
chore: all chores/tests will be excluded from the CHANGELOG

Please add a description with these things:
1. A good description explaining the problem and what you changed.
2. If it fixes any GitHub issues, say "Fixes #GitHubIssue".
3. If it corresponds to a Jira issue, say "Fixes DGRAPH-###".
4. If this is a breaking change, please put "[Breaking]" in the title. In the description, please put a note with exactly who these changes are breaking for.
-->

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/dgraph/7435)
<!-- Reviewable:end -->